refactor: centralize renderer draw and cleanup mesa checks
