+++ src/cef/graphics/gpu/cef_renderergpulinuxmesa.h
@@ -1,8 +1,7 @@
 #pragma once
 
 #include "../cef_renderer.h"
-#include "gpu_helper.h"
-#include <framework/graphics/texture.h>
+#include "gpuhelper.h"
 
 #if defined(USE_CEF) && defined(__linux__)
 #include <GL/gl.h>
@@ -24,12 +23,9 @@ public:
     void onPaint(const void* buffer, int width, int height,
                  const CefRenderHandler::RectList& dirtyRects) override;
     void onAcceleratedPaint(const CefAcceleratedPaintInfo& info) override;
-    void draw(Fw::DrawPane drawPane) override;
     bool isSupported() const override;
 
 private:
-    TexturePtr m_cefTexture;
-    bool m_textureCreated;
     int m_lastWidth;
     int m_lastHeight;
     mutable bool m_checkedSupport;
