+++ src/cef/graphics/gpu/cef_renderergpulinuxnonmesa.h
@@ -1,23 +1,18 @@
 #pragma once
 
 #include "../cef_renderer.h"
-#include "gpu_helper.h"
-#include <framework/graphics/texture.h>
+#include "gpuhelper.h"
 
-class CefRendererGPUWin : public CefRenderer
+class CefRendererGPULinuxNonMesa : public CefRenderer
 {
 public:
-    explicit CefRendererGPUWin(UICEFWebView& view);
+    explicit CefRendererGPULinuxNonMesa(UICEFWebView& view);
     void onPaint(const void* buffer, int width, int height,
                  const CefRenderHandler::RectList& dirtyRects) override;
     void onAcceleratedPaint(const CefAcceleratedPaintInfo& info) override;
-    void draw(Fw::DrawPane drawPane) override;
     bool isSupported() const override;
 
 private:
-    TexturePtr m_cefTexture;
-    bool m_textureCreated;
     int m_lastWidth;
     int m_lastHeight;
 };
-
