+++ src/cef/graphics/gpu/cef_renderergpuwin.h
@@ -1,22 +1,19 @@
 #pragma once
 
 #include "../cef_renderer.h"
-#include "gpu_helper.h"
-#include <framework/graphics/texture.h>
+#include "gpuhelper.h"
 
-class CefRendererGPULinuxNonMesa : public CefRenderer
+class CefRendererGPUWin : public CefRenderer
 {
 public:
-    explicit CefRendererGPULinuxNonMesa(UICEFWebView& view);
+    explicit CefRendererGPUWin(UICEFWebView& view);
     void onPaint(const void* buffer, int width, int height,
                  const CefRenderHandler::RectList& dirtyRects) override;
     void onAcceleratedPaint(const CefAcceleratedPaintInfo& info) override;
-    void draw(Fw::DrawPane drawPane) override;
     bool isSupported() const override;
 
 private:
-    TexturePtr m_cefTexture;
-    bool m_textureCreated;
     int m_lastWidth;
     int m_lastHeight;
 };
+
